setup/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4